PHOENIX-2014-T-release-v3/PHOENIX-2014-T/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
imageio>=2.31.0
Pillow>=10.0.0
pyyaml>=6.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0